+++ JanitoEditorExtrasPackage/Runtime/Attributes/ReadOnlyAttribute.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace Janito.EditorExtras
+{
+    public class ReadOnlyAttribute : PropertyAttribute { }
+}